+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 - Fixed an error that could appear if a campaign type's or mailing list type's site was deleted.
 - Fixed a bug that was causing errors when using Postgres ([#128](https://github.com/putyourlightson/craft-campaign/issues/128)).
+- Fixed a bug in which a contact could be unsubscribed from a mailing list other than the one they opted to unsubscribe from when clicking the unsubscribe link ([#132](https://github.com/putyourlightson/craft-campaign/issues/132)).
 
 ## 1.14.5 - 2020-04-03
 ### Fixed